Move RunServer ⇒ Server::Run
