+++ lib/http/commands/controls.rb
@@ -8,6 +8,6 @@ require 'http/commands/controls/messages'
 require 'http/commands/controls/messages/requests'
 require 'http/commands/controls/messages/resources'
 require 'http/commands/controls/messages/responses'
-require 'http/commands/controls/run_server'
 require 'http/commands/controls/server'
+require 'http/commands/controls/server/run'
 require 'http/commands/controls/uri'
